README
Write a program that generates a set of samples of a sine wave. The parameters are as follows: 
Frequency = last 2 digits of your index number [Hz] (if it is ‘00’ then use ‘01’), Sampling Frequency = 
48 [kHz], Acquisition time = 2 [s], Amplitude = 2. Hint – use numpy and matplotlib libraries. It should 
be realized as a class which stores the sine wave, has a method to generate sine wave samples with 
chosen parameters, plotting the sine wave as well as returning samples of downsampled wave. Plot 
both sets on one figure for comparison