Add files via upload
Write a program with the same functionality as in List 1 task 6. It should be realized as a ‘Player’ 
class, where each of the player is a separate object. Inside it stores the score of the player and has 
method for calculating the score of a given word. Make a simple text UI to enhance the program's 
accessibility for players